requirements.txt
mediapipe
PyYAML
loguru
torch
pandas
scikit-learn